pyproject.toml
[tool.poetry]
name = "python-challenge"
version = "0.1.0"
description = ""
authors = ["stealthmate <stealthmate1233@gmail.com>"]
readme = "README.md"
packages = [{include = "python_challenge"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pillow = "^10.2.0"
numpy = "^1.26.4"
matplotlib = "^3.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
